[advertisement] refactor : DTO 위치 변경
- 응용 계층에서 인프라스트럭처 계층의 DTO를 직접적으로 참조하는 방식을 수정하였음
